refactor: 결제 완료시 예약 상태 변경
