+++ src/test/java/roomescape/reservation/domain/ReservationTest.java
@@ -2,7 +2,7 @@ package roomescape.reservation.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
-import static roomescape.reservation.domain.ReservationStatus.*;
+import static roomescape.reservation.domain.PaymentStatus.*;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
